refactor handler code
